progress commit

+++ Real-Time Event Ticketing System/backend/src/main/java/lk/ac/iit/backend/repository/VendorRepository.java
@@ -3,5 +3,5 @@ package lk.ac.iit.backend.repository;
 import lk.ac.iit.backend.model.Vendor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VendorRepository extends JpaRepository<Vendor, Long> {
+public interface VendorRepository extends JpaRepository<Vendor, Integer> {
 }
